Add "addOnCreate" option prop
I had to make this change for my project because I'm processing the created data before attaching it to the feature group. By supporting `addOnCreate` option it looks intuitive, when set to false, to asynchronously handle the created data through `onCreated` event (even not adding the layer if the backend returns error).